requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
